requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
anthropic==0.18.0
chromadb==0.4.22
sentence-transformers==2.3.1
python-dotenv==1.0.0
pytest==7.4.4
